errors were fixed
